requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
isodate==0.6.0
pkg-resources==0.0.0
pyparsing==2.4.7
python-telegram-bot==13.7
pytz==2021.1
rdflib==5.0.0
six==1.16.0
SPARQLWrapper==1.8.5
tornado==6.1
tzlocal==2.1
